Drawing test.
This removes game launch, and updates the drawing code for testing.

The plasma style effect is applied in C# here to draw the Texture2D for the Quad.

This is running and works in VR, but is slowish and low res because of the SetPixel use.

Took forever to figure out that Colors are 0..1, not 0..255