Setup to return DX9 Surface to C#
The c++ code now creates a IDirect3DSurface9 as a RenderTarget when the CreateDevice happens.  This RenderTarget matches the backbuffer created.

During Present, the backbuffer is copied to that RenderTarget with StretchRect.  No errors are reported for any of those functions.

Then, we call the CustomAPI of GetGameSurface, which returns the address of the IDirect3DSurface9.

I verified that the CustomAPI is working, and returns the actual address of the object on the video card memory.