Snapshot structure. Not working.
This is very much close to what we'd want for structure.  Moved the fetch of vtbl addresses into a different C file altogether, and changed API for it.

This handles the Ex problem with header file, and has all the pieces we'd expect to need.

Compiles, but crashes at run.