Regenerate .sln and .csproj files.
These are now used solely by the Unity project, so regenerate it to match whatever it wants, by deleting the prior files.

This changed the GUID of the project and names.

We want to keep the .sln file included in source control, because this subproject will be added to the overarching Katanga project.