Setup and running.
This now runs with the new setup, new configuration of project folders.

Moved the destination of DeviarePlugin to be \UnityScreenApp\Assets\Plugins\x32 to more closely match the expected Unity structure.

Decided not to have Debug/Release target folders, because generally it's unimportant.  Only x64/x86 matter.

The DeviarePlugin will not be added to Unity project because it is not called from Unity directly, but can still live in the same folder as the UnityNativePlugin.