Holy shit, it works.
This is still a bit of a mess, but damn if it doesn't work!

Getting bits from the 'textures' app, and drawing them into the Unity big screen view.

Stretched to match the screen, and seems to be upside down or something, because it rotates the opposite of the game.

But damn....  Got bits from a 32 bit DX9 app into a 64 bit DX11 Unity app.

Also did not need a second copy, once I got the proper resource back from the native plugin of a SRV, I passed that directly to Unity to use.  So, StretchRect copy from the original game/app to a shared surface that is directly used in Unity. No second copy, no going to main memory.