Get TheBall running.
This is pretty epic- managed to get TheBall to draw in the Unity big screen TV.  It's upside down, but it follows the game window, and both are working simultaneously.

Required adding 3 more hooks for different calls the game makes, where the change to DX9Ex changed semantics.  In particular, we cannot let the game use Managed_Pool, it must now be Default_Pool.  This leads to debug layer complaints that Lock cannot work on non-dynamic surfaces, so those are also marked Dynamic.

This has been working, and runs at full performance.  75Hz in Unity, 60Hz/120 in TheBall in 3D, while using 45% of the GPU.  Good start.