Invert TV screen. Runs in VR.
This changes the TV screen scale to a negative number, which inverts that Quad. It was inverted by Unity to make it match OpenGL, this de-inverts to make it match DX, which is where the bits come from.

Works!  Runs in VR!

No 3D, but it runs at good speed.

Couple of weird problems, if the game is full screen to start it will crash.  Needs to be run from BuildAndRun, just the direct build will crash in Deviare.