Always return D9Ex object.
Change startup launch to always return the D9Ex object, and remove the duplicate from InProc, and use QueryInterface upcasting to fetch the needed D9Ex object.

This is logically superior, because the game will then always use the D9Ex factory.

Also set Unity to run in background, so that updating happens, even if game is in foreground.

Improve the logging, to include the types of data created for CreateTexture et. al.