Fix crash on full builds.
This was weird.  Deviare would crash if the NativeDeviare.dll could not be found, instead of returning an error.

Path was wrong for full builds.  For normal builds, the Environment woudl be right, but for full standalone builds, it would not.

So this changes the paths and the dll names to make it better for the full builds.  We only have a Plugins folder now, not x86 or x64 variant, and move the architecture into the name64.dll type.

Now is running completely as standalone, as well as in Unity during debugging.