Add nvapi for stereo reverse blit.
This includes the headers, inits nvapi, and uses the NvAPI_Stereo_ReverseStereoBlitControl to fetch the stereo bits from the game.

The shared surface is 2x the game size, and seems to work.  No errors are reported.

TV display is not stereo though, probably because Unity does not know to treat it as stereo.