Change to VR mode now.
Close enough that we need to see stereo happening or not, and VR mode changes a few things, including adding the single-pass stereo which we want to use.

There is the option of None for a VR headset now, which allows us to keep it in VR mode, and skip drawing to physical headset for debugging.