Add Sand under feet, and ms counter.
This adds the two items to the project file.  The sand is to give you something to stand upon, to make it less feeling like falling.

The ms timer uses the SDK values to get a specific time used in the SDK to draw the full scene.  Does not include game time.  Does not work when using Oculus SDK.