Fully working!  SinglePassStereo.
This is working all the way from game to VR, all as designed, and with only two StretchRect copies, and one shader copy from the game bits.

Image is now stereo in the virtual 3D TV, and looks exactly right.  Eye swap happens during the shader, and doesn't cost anything extra.

Last step was to add the sbsShader here, which replace the normal shader for the Quad Material.  This copies from the offscreen 2x width stereo texture, to the VR screen.

It does the work required anyway, which is to show the bits in the VR world, and move them to their right spot.  But, also fetches the correct eye, and uses the unity_StereoEyeIndex variable in the shader to know which eye is active.

Needs some cleanup, but this is fully functional.

All the way from game to VR.  Miracle!