Add copy thread in game.
This makes it multi threaded in the game app, where the 2nd copy from the stereo bits to shared surface is done on a different thread.

This should free up the game to immediately continue after a single StretchRect.

Not sure this is best, getting dropped Compositor frames, that seem like probably pipeline stalls because of beat frequency interference.  GPU usage is about 50%, so plenty of headroom.