Switch all Ex use to normal DX9.
Turns out that only the DX9Ex factory needs to be Ex, the normal calls will still create Ex objects, but none of the interfaces need to change.

This cleans up the code to use normal interfaces instead of the Ex variants.

Also solves the problem of full screen vs. windowed.  Since we create a standard Direct9Device now, we don't need to modify the game's expected window layout.

Works in The Ball to be full screen at launch now.