Add SelectGameDialog to allow game selection.
If LeftCtrl is down, we'll put up dialog to select a different game.

Actually kind of tricky from the Mono version of C#, and no .Net.  Needed a special marshalling command to get the wide strings to the C++ side where it could just call an old school Win32 GetOpenFileName.

If used in Editor, it will freak out, because of swapping out.  It breaks the working directory.