Checkpoint for performance issue.
Flashing F in the Rift, dropped frames.  Analysis shows that it's TheBall taking higher priority because it's window is frontmost, and denying the end of frame to the critical VR in Katanga.

Tried multiple ways to solve this, nothing works. NVidia and Microsoft provide no way to change GPU priority.

This code includes some attempts to fix, but will be removed.  It's a relatively minor issue in TheBall, need to try other games.

GPUView shows that there is an 8ms hunk of TheBall code that runs on GPU, and when that overlaps endofframe, it drops the frame in Katanga.

No good answer at present, will have to come back to it.