Experimental sync test.
This doesn't work well, but is an interesting idea worth saving in history.

The idea was to try to move TheBall sync to earlier in the frame, so that the time when the game pre-empts VR would be different.

Doesn't work, still stalls the VR frame.

Does not seem to be any developer way to change the priorties on the GPU.