Activate the mutex for TriggerEvent.
Changes the creation of the mutex to be used for syncing between the game and Katanga.  Whenever Katanga starts a new frame at FixedUpdate, it will fire the TriggerEvent via the Deviare CallCustomAPI.

The game code will be stalled right after their Present, waiting for this event.

Thus, as along as the game is running faster than 90hz, it will remain in sync with drawing in VR.

This helps lower the flashing F seen in VR, where there is contention over the GPU use, and the game wins because it is in the foreground.  Probably not quite right, needs further profiling, but it is fully functional.

Game is actually running at solid 50Hz now, which I can't explain.