Fix a constant LoadLibrary/FreeLibrary sequence.
Calling CustomAPI was donig a LoadLibrary for every call, then a FreeLibrary at that exit. Waste of resources, and noise in debugging.

After sifting the tea leaves, this wound up being changing the original LoadCustomDLL to FALSE for the unload on exit. Exit in this case counted as no other calls, not process exit.

So, it remains loaded now in the game process.