Switch sync to WaitForEndOfFrame.
This moves the syncwait to the end of the VR frame.  And after Present in the game itself.

Using this spot means that our VR code runs quickly to completion, as it's less than 3ms to do our simple scene.  Then it waits for the Oculus vSync, and only has to present its data. Right at that moment, we free up the game to start a new frame, and the initial part of a new frame is CPU based physics and layout, no GPU, so there seems to be no contention over GPU resources this way.

Fixed the flashing F completely.  No F no C, no dropped frames, game itself runs at 90Hz.