Add DrawIndexedPrimitiveUP hook for timing.
Dropping frames in VR app, because we cannot have precedence over the game.   This adds a hook for a call that TheBall makes all the time, giving us the opportunity to stall the game when it's time for us to Present.

Doesn't seem to work quite right, because the timing for Unity games is too low resolution, at one per frame, not ms as the call describes.

Includes an improved on/off setup from the DrawSBS.cs where it will set or reset the Trigger event in the C++ side using a Deviare call parameter.