Timing test case with microsecond accuracy.
This is some interesting tidbits, including a microsecond accurate timing tool, which is vastly better than the normal 15ms minimums.

Doesn't work.  The idea was to stall TheBall by hooking it's drawing calls, and making them wait for the next frame when Katanga might be in trouble of dropping a frame.  Setting it to even a 3ms max for the game to run did not work.

Using GPUView I think it's working correctly, there are gaps in TheBall while it's drawing, but it doesn't solve the problem of dropped frames.

This is a real sick hack anyway, and clearly not the way to go here.  Worthy only for test code, and the timing samplers.