Create ReleaseBuild script.
This is setup as a menu option in the Unity editor, so that a build can be created at any time, without needing to use the normal Build menu manually.

Scripted.  Creates the full Release build, and cleanly deletes prior one.

This also allows copying needed files to the output, like the deviare32/64.db files for SpyMgr.Init to succeed.