Add notes regarding RegFree COM.
Doesn't work.  Cannot get the manifest approach to use a different directory.  API is clunky, documentation is terrible. No experiments could discern a way to add RegFree.

Skipping this for now at least.  Not a huge deal, because we are very likely to have an install program anyway, which can simply call regsrv32 at install time.