Experiment to get around old Helixmod.
Older HelixMods do not support the Create9Ex call, so it won't run.  Newer versions do, but changed all the syntax so it would require modifying old fixes.  Undesirable.

This attempts to call the System32 version of Create9Ex directly, and avoid the missing call.

Mostly works, does not crash, but does not solve the problem of HelixMod not hooking into the game (Psychonauts).

Rather than beat this dead horse, this is a save snapshot, and moving to look at doing encode/decode instead of surface sharing.