Add standalone NvCodec code.
This is interface code to the NVAPI for encode/decode of video.  Used in the sample projects to more easily setup and tear down the codec pipeline.

This is part of experimenting with using video encode/decode instead of DX9ex surface sharing to get the video from game to Unity VR app.  Nvidia has much more visibility into the surfaces, and the codec can be lossless.  Only real question is whether it's fast enough.