First stab at setting up nvcodec.
Unbelievably complicated, with parameters and options out the ying yang.  Very poor sample code that demonstrates every feature under the sun, but nothing showing the simplest path.

This sets up the basic structures we think we need, and a starting point for encoding the game buffer after a StretchRect to the nvcodec InputBuffer.

Surely will generate errors, but is a good checkpoint.  Compiles, but that's it.