Some experiments for NvEncRegisterResource.
NvEncRegisterResource is returning null for the resource, while also returning no error.  Lame.  Trying experiments to see if different surface types would work.  No luck.

Seems to require OffscreenPlainSurface in order to succeed, but that surface cannot be StretchRect or any way to get bits from backbuffer to it.

Moving to trying to use sample code example of DXVA2.