Cleanup comments regarding Create9Ex.
This is a must for cuda as well.  The rest of the objects don't have to be Ex type, like Device, but the factory object must be Ex type or cuda will throw an error.

This commit just cleans up the comments and rips out other speculative code, and settles upon the Direct3DCreate9Ex call, which is returned.