First pass complete code using cudaGraphicsResource.
Compiles, links, not tested.

Renamed all functions to reference cudaGraphicsResource instead of HANDLE.  Rewrite all comments referencing prior approach.

This includes changing the names and return values for the embedded IPC call that Nektra manages.

Adds code to the DX11 UnityNativePlugin to create the Texture2D resource ,then the ShaderResourceView that Unity needs from that.  Includes the ArrayToArray cuda copy code that was developed in the sample application.