Switch to CudaDriver API from CudaRuntime.
CudaRuntime doesn't support cudaContext managment, so we are switching to the driver level version which does.  These are a bit harder to use, but this sequence was sussed out in a sample app, and seems to work here.

We get no exceptions or errors, and the g_cudaStereoResource has a valid result.

Does not fully work here, the DX11 side needs modification to set the cudaContext.