Save mid-stream Cuda copy changes.
Moving back to Surface Sharing, so this Cuda copy needs to go on hold.  Wasn't going to work here anyway, because it does not appear to be possible to share a CudaContext across different processes.  Would most likely work with an embedded VR environment, but not with the Unity app.

This will be archived in a named branch.