Add D3D11CreateDeviceAndSwapChain hook.
Switching to Alien Isolation for testing, even though it is x32 game.  Known good test case, without DRM complexity.

It uses D3D11CreateDeviceAndSwapChain instead, so setup to use that instead of CreateDXGIFactory.

This then gets SwapChain directly, and we fetch the parameter from the stack using Nektra and create the Present hook directly.

Compiles, but is untested.