Add the GetSharedHandle for the stereo copy texture.
Makes the Texture2D that is double width a shared handle, which can then be fetched in Unity plugin.

Also set up the ShowOnGame debug display to see if we are getting stereo bits out of the game, and into the gGameTexture properly.  Clunky, because no stretch in DX11, so just offset to see if right side is black or bits.

We are getting stereo bits, as long as we are in Discover mode.  And presumably full screen exclusive.

So the gGameTexture is shared, and properly full of stereo bits.  Not being seen in Unity app though, still all black.