Clean up HookPresent for single routine.
Was duplicated to CreateSwapChain, and that made no sense, so this moves all the HookPresent code into that routine, and calls it from CreateSwapChain.

HookPresent can be called directly from CreateDeviceAndSwapChain, which is why it is standalone.

This works in Witcher3, which calls CreateDevice, CreateDXGIFactory, then CreateSwapChain.

Also of note- when creating the shared stereo Texture2D, sets the BindFlags to include D3D11_BIND_SHADER_RESOURCE, because that is required by the Unity layer when creating it's ShaderResourceView.  It's otherwise a 2x width exact duplicate of the backbuffer of the game.