Holy shit, it works!
This is fully functional with DX11 sharing, using Witcher3 as the test case.  Getting full stereo image all the way to the virtual TV.

Performance is stellar.  At 1080p for game, it's pulling a full 90 for the Unity app, and 60 for the game.  No drops.

This commit is small cleanup of _DEBUG ifdefs for the Release x64 build.