Filter FatalExit to just exceptions.
Not needed for every log message, just the Exceptions thank you very much.

Tested this, and it does work.  If DLL could not be loaded got "Could not load NativePlugin DLL".  If Deviare was not registered, got ComExc, which is from Deviare.

Seems like a perfectly good way to report fatal errors without the tedium of logging.