Use game width/height/format for Unity texture.
Passes back the full game information to the Unity C# side, so that it's CreateExternalTexture can use the exact same size and format as the game itself.

This completes the chain from the game backbuffer to that displayed by Unity, so that game resolution should make it all the way to the virtual screen and look better.

Was hard-coded to 1600x900, now it's whatever the game is using.

Works still, no errors.  Not positive the Format is correct.