Move game selection dialog to Awake, to solve bad directory crash.
Unity will sometimes crash at the game dialog selection as the working directory gets modified by the dialog.  Shouldn't, but it does.

This adds the OFN_NOCHANGEDIR to the dialog flags as well, which might have helped.

Moved the dialog selection into Awake, because it makes more sense there than Start.  It is set to always ask for game now as the first step.

Seems to solve the crash while using the editor.