Handle both x32 and x64 games, with proper plugin load.
This determines from the _gameProcess, whether it's x64 or not, and loads the proper plugin name.  Rather than relying upon Deviare to produce decent error results.

If an error is seen, we throw an informative exception.

Made GetLastErrorCode to fetch the Deviare error in case it happens, the real one, not their fake return results.

Moved the error exception handling to bottom of file as a better spot for rare events, keeping Awake and Start at the top.

Works correctly to launch x32 or x64 games.