Change TextureFormat to always be RGBA32.
It's not clear what is best here, the game may not be using a texture format that maps nicely to the Unity versions.

In general we expect games to use DXGI_FORMAT_R8G8B8A8_UNORM, so if it's not that or the sRGB variant, we'll throw an exception for now, so we can see when it happens.

Otherwise, these map nicely to the CreateExternalTexture and should look good.