Add --game-path cmd line parameter to specify game to launch.
In Awake, checks cmd-line for --game-path and if found uses that path to launch the game.  Makes it possible to use shortcuts, and should work with 3DFM.

If no path is specified in params, uses the StandardFile box to let them choose. If they hit cancel there, throws an exception for no game specified.