Add launching InfoText for user experience.
Pain in the ass to get Unity to draw text as a sub object of the Quad.  Wanted to use the 'TV' as the place for text, as it seems like the obvious spot to look.

Added a Canvas, then Text object to the model, and got all the million scale and position locations worked out.

Added setting and clearing the text in the SBS.cs code, so that the text is removed once the game connects.

Looks pretty good.

Also shrank the floor to 5x5 from 10x10 as 10 meters is a bit much.  Added a 1 meter cube for sizing purposes to be sure it looks right scale.