Merge in archive_unity changes for DX9.
This is using checkout archive_unity * for these files, so strictly speaking that's not a merge, but we don't need everything in that branch, just tip of tree of DX9 support.

This merges in those changes manually into the files needed, with small tweaks to get it to compile.

This is the basic structure, compiles but not expected to run.