Clean up variable handling, moving globals to DeviarePlugin.cpp.
With the addition of the InProc_DX9, we get conflicts on the globals being defined in different compilation units.

This change moves all the global variables into the DeviarePlugin.cpp as the owner of the dll.  They are declared extern in the DeviarePlugin.h file, so that they can be referenced in InProc_* files, but only defined in the DeviarePlugin.cpp file.

This makes a single copy of every variable and the important GetSharedHandle call that returns the gGameSharedHandle to the VR side.

Runs and works correctly for DX11 side again.