Fix couple of small bugs.
Debug builds would break on the GetAt calls,so set those to use the &param.p to get pointer and avoid the operator check.

DrawSBS would crash when turning off the left/right screens, so set that up to handle null returns if they are off.  Will work normally if they are on.