Fix bug with launching UE4 games, (IDXGIDevice used).
Just like in 3Dmigoto, got stung by their bad usage of IDXGIDevice passed in to CreateSwapChain.  This makes no sense to use that device in a 3D game, but they do it anyway.

Fixed by not using what they pass in, just use the swapchain created, and fetch the ID3D11Device directly from that with GetDevice.  This is more reliable in any case.

Worked to fix the problem with UE4 games, so now Little Nightmares was tested as working.