Allow any color format through. UE4 seems to use DXGI_FORMAT_R10G10B10A2_UNORM.
No good mapping from DXGI_FORMAT_R10G10B10A2_UNORM to a Unity TextureFormat.  There is a mapping to RenderTexture, but that's not what we are using.

Does not seem to cause any problem to just force Unity to use RGBA32.  No idea why, because the color mapping should be all over the map from 10 bit cut down and mixed to 8 bit.

In any case, letting it through for now, no Exception, because an Exception in a UE4 game causes it to hang, not report it.