Add ResizeBuffers hook and handling for UE4 games.
This is also for any time the user will set the resolution in-game, but UE4 does this on launch, and thus needs a better handler.  It's a must to be able to handle resizing in any case.

This hooks the IDXGISwapChain->ResizeBuffers call now too, and during that call CreateSharedTexture, which is an abstracted call that was previously in HookPresent.  Called from both places now, at first sight, and after resolution change.

In the DrawSBS.cs, removed the coroutine waiting for the gGameHandle to show up, as Update is called while it is waiting anyway.  Moved the logic to a new routine PollForSharedSurface, and reset the Quad texture whenever it changes.

This also means that at any given Update, it can switch the shared buffer to a new resolution, and thus should be transparent to changing resolution.

Tested in TuringTest, and it worked correctly.