Slightly improve resolution change handling via coroutine.
Putting change into Update doesn't really make sense, because it can stall drawing.  So can coroutines since they run on the main thread, but Unity gets to decide when to run it.

Put it back to being a coroutine with a 200ms yield, so that it checks every 200ms for new resolution to be setup.

This works, but gets a crash after doing it like 3 times, which suggests maybe Releasing something that should not be.

It helps with screen stall, stuck to face VR when changing resolution, but not when game is in full screen.  Don't think we can fix that, even with true Thread, because it's resetting the GPU.