Add resolution handling for DX9 via Reset hook.
IDirect3DDevice9::Reset needs to be hooked to handle changing game resolution for DX9 variant.

After Reset, this fetches the new backbuffer settings, and fires off the rebuild of the drawing chain.

Doesn't always work, in The Ball, a UE3 game, if you grow the dimensions it's called, but if you shrink, Reset is not called.