Delete all .meta files for Project plugins.
The .meta files are a versioning mechanism that Unity uses to connect stuff like Materials to a Quad.  They are automatically generated for any of our Plugin files.

However, we do not need this versioning mechanism for our generated files, and like always we really do not want generated stuff checked-in.

So, deleting all of these .meta files, but only for our Plugin files that have no other connection to Unity pieces.  They will be auto-regenerated, but the .gitignore already says to ignore everything in Plugins.

This will solve the conflicts and changeslists for these whenever we rebuild the C++ plugins.  It will introduce warnings at Unity launch, which we will ignore.