Put PollForSharedSurface back into Update.
Switching yet again, but the idea that coroutines were more efficient is wrong.  They are the same as Update, just a different location right afterwards.

No advantage for this use, and adds unneeded complexity, so we are putting it back into being called every frame at Update.

The other reason to revert this is because we are getting crashes when changing resolution in-game, and we need to try harder to avoid using disposed textures.  And having a coroutine made that more complex.