Set gGameSharedHandle to Null whenever resolution changes.
Have seen crashes during resolution changes, so this attempts to resolve that by handling the drawing texture that the VR side uses more carefully.

Whenever we get a DX9->Reset or DX11->ResizeBuffers, we will immediately set the gGameSharedHandle to null, to notify the VR side to stop drawing.  It will notice that null during polling, and immediately switch out the drawing texture back to the grey screen from launch.

Only after the shared handle is rebuilt will we return to drawing using a new texture for the screen.

This should hopefully prevent Unity from *ever* drawing using a lost image/texture because it will reset this during Update, before drawing the next frame.

This change handles the difference between DX9 and DX11.  In DX11 the textures and rendertargets are still alive, in DX9 they have already been destroyed.

Not positive this will fix the crashes, but this is a much more solid approach, unless we ultimately want to keep the sides in perfect sync.