Rewrite hook handling for Direct3DCreate9.
Turned out that the previous approach was actually unhooking the HelixMod d3d9.dll, which disabled it from fixing the graphics.

This approach now moves the hook setup code into DeviarePlugin at the OnLoad call of the native plugin.  That's as early as it gets, and before the app gets to run, so we can set the hook.

The hook needs to be done on the System32 version of d3d9.dll, so we specifically fetch that dll and add the hook to the procedure there.

This has the net effect of inserting our Direct3DCreate9 hook in between HelixMod and System32 versions, so we can force the return of the IDirect3D9Ex object.