Switch color space to Linear from Gamma.
Helifax reported that TheSurge looked too dark, and the color matching in the code seems to pass the proper color format all the way to the VR screen.

For some reason, the Color Space in Unity was set to Gamma, instead of Linear.  Linear is the currently best choice for color mapping.

I seem to vaguely remember that I might have set it to gamma to support single-pass stereo, but switching it to Linear works fine and solves the problem.

With Linear for the Unity environment, TheSurge looks much, much better.