Add 3D Vision DirectMode handling.
This hooks the nvapi.dll, of QueryInterface, and watches for any call to the SetDriverMode call, which is used to set DirectMode.

If we see this call, we'll set the gDirectMode flag, and that will then change the drawing behavior in Present, to use SetActiveEye to fetch each eye buffer.

Seems to work correctly in Tomb Raider 2013, a game which does not support any SBS output.