Hook NvAPI_Stereo_SetDriverMode instead of QueryInterface.
There is a bug with In-Proc that generates bad code when hooking QueryInterface.  It does not update a RIP relative address, and thus will crash for x64 hook of nvapi64.dll.

This changes the code to just directly hook NvAPI_Stereo_SetDriverMode directly, by fetching it's address by calling QueryInterface at OnLoad time.

This is actually a bit better anyway, more to the point of what we need, less code.  Took a full day to debug though.