Remove DX9 SetSurfaceCreationMode.
Not sure this did anything, and can't find any reason why I added it.  Commenting it out for now, as something that is probably not the right idea to do in the game space.

Also remove the SetGPUThreadPriority, which was tested as doing nothing. Goal was to improve VR performance to avoid dropped frames- does nothing.