Update to latest version of Nektra In-Proc.
This includes a bug fix for the bug I found in In-Proc, where it would not properly fix a RIP Test instruction.  Nektra kindly fixed this bug quickly, and this is a newly compiled version from that source.

This fixes the problem seen with x64 DirectMode games.

However, that code is no longer tip of tree, because I'd already added a workaround that is better code.