Enable MSAA in Camera.  Add Low setting.
AA was disabled completely because the camera disallowed that.  This reenables it in the camera, and adds a Low setting which people can switch to, to disable AA altogether.

Having options is better than having it disabled.  Setting Default to Low however for performance.  AA should be unnecessary in Katanga, as aliasing the screen contents is probably not a good idea.