Fix Gamma problem once and for all.
Took awhile to figure out, but the problem was that TheSurge would specify  an sRGB backbuffer, which we'd copy.  That would double up the gamma correction, once from the game, and once for the HMD, making it too dark.

Setting Linear in Unity was not the right choice, because that made normal games too bright, with no gamma correction on their output.

So, this commit switches us back to Gamma for Unity itself, which will work in 90% of the games.  And adds a specific handler for the sRGB formats when we CreateSharedTexture in the game, and force it back to Linear for the game.

Tested as working in TheSurge.