Quality improvements, add Quality +/- keys.
Sets the backdrop to not be a skybox, but rather a solid black color.  The floor is still visible, so this seems like a good compromise for a quick fix.

Sets the default Quality Level to Ultra instead of Low, with 8x MSAA.  This fixes most of the shimmer that we've seen. Not sure Ultra is the right choice, 4x seems OK.  8x costs maybe 2 fps, so it's not really cheap, but Low=off is not the right choice.

Adds keyboard handling for [ ] to decrease/increase the Quality on the fly, so people can experiment.