Switch to MessageBox instead of Exceptions for Unity plugin too.
The x64 UnityPlugin does not seem to properly report error seen there that are thrown as exceptions.  Should be part of the Unity runtime, but their handler doesn't catch them, and causes hangs.

Switch to MessageBox as a FatalExit handler instead, so we can see any of these errors.

Using exceptions is too problematic in this complicated runtime environment.