Add QualityLevel text on the floor.
It's hard to tell what state the QualityLevel is in, so this makes a text object on the floor that will write out the current quality level.

Maybe too subtle, but at least something you can see directly while testing different levels.

This moves the DrawSBS shader Script object onto the Camera from the Quad, because it makes more sense there.  And now that there are two text items, it needs to be different.