Fix recentering.  Introduced bug.
Thought that it made sense to save and restore around the rencenter, but the way it actually works is to set the tracking space to Stationary.

I'm not sure we actually want to do that, it seems pretty bold to change the user's space type, but it's currently the only way to recenter the HMD.

In normal room scale VR, the HMD is not reset as it knows the entire space.