Fix DrawSBS to match new location of script on Main Camera.
Doing the Renderer lookup was failing, because the script is now on the camera.  Rather than even do the lookup, we'll just use this technique of creating a global in the DrawSBS script and connect it via the Main Camera.

This now just directly links, like the InfoText and QualityText items.  Renamed to Screen from Quad.

Lookup works and runs again.