Add QualityLevel text, and allow anisotropic to be set.
Expand the QualityText string to include current MSAA and Anisotropic settings.

Enable "\" key to enable/disable Anistropic filtering.

Sets Anisotropic to on by default for all levels, under the idea that the screen can be at an angle to the viewer, so we should clean up the image.

Can't actually tell much difference, but at least you can see what the settings are.