Add DXGI_FORMAT_R10G10B10A2_UNORM=24
This is deep black mode, 10 bits per channel.  Doesn't really make sense to generate an error, it seems to work fine.

Alien Isolation is a good test case, because its deep-color mode is =24, and you can enable or disable it.  Cannot see any difference in headset, but did not look for dark black areas.

This is also used in Call of Cthulu and ME:Andromeda, so this change just lets them work however they look, without reporting on it.