Experimental branch using SteamVR Unity Plugin.
Installed and setup for using the SteamVR Unity plugin, which provides a lot of default features that we very likely need, like floating controllers, and some UI connections.

Set up to use their prefab CameraRig as the main camera, and delete old one.  Reconnected SBSShader to this new camera.

Not sure this is setup right, hence the branch.  Does seem to run however and does not break the 3D.

RecenterHMD is broken now though.  Floor is also wrong height again.