Switch to using the Player style of SteamVR, instead of Simple.
Simple doesn't support text hints, and Player already has a mechanism for that.  Probably a bit heavyweight, certainly includes a lot of unnecessary baggage, but it's probably better to have the full deal for quicker progress.

This changes to using the Player as the prefab for drawing in VR, and the SBSshader is attached to the VRCamera.