Setup all pieces to change Menu button on controller.
This creates all the bindings necessary to setup the ToggleHintsAction for the Menu button.  It's a bit convoluted, but it's working here.

This just creates C# file of ControllerHints.cs to enable/disable the hints mechanism for the buttons.  It works to show all the current bindings from the interaction sample.

This will be useful once we connect screensizing and other functions to the controllers.

Runs in VR, still getting 3D TV, now with controllers floating in space with active Menu buttons.