Add ScreenSizingAction to grow/shrink screen.
This is tied to the left controller for scrolling, and as you scroll it will grow or shrink the screen in place.  Visually it looks similar to distance, but this changes the scale of the transform.

Tricky part was that Y axis is a negative number for Unity, so creating Vector3 needed a negative deltaY.

Works to grow/shrink.