Add Coroutine to make it auto-repeat.
Switch back ChangeListener to see the down, then up of the d-pad button, but have it fire off a StartCoroutine and StopCoroutine for each state change.

The coroutine will pause and be rerun after WaitForSeconds, and thus auto-repeat until they let up on the button.

A bit of a challenge to get all the ducks in a row to get this working, but auto-repeat seems best.