Add remaining actions, Bigger/Smaller, Higher/Lower.
These are the Left hand operations, for d-pad on left hand. Horizontal is smaller/bigger.  Vertical is higher/lower.

Connect up the actions through the ScreenDistance.cs object references.  Each is a specific action as this is a d-pad click.  The coroutine is just to give it a repeating action.

Seems to work pretty well.  No errors or warnings while running.  Very smooth operation at 10cm every 20ms.