Change default of rendermodel to standard controller.
Seems like this is the actual answer to changing the render model while still using the Hand script.  We want Hand for use in hints.

This changes the default in the Right/LeftRenderModel prefab, so that it uses the default controller and not the gloved hand.

Tested and is working.  No errors or warnings.