Remove the UnityScreenApp project from Katanga.sln.
This is partly broken because of the extensive changes to the Unity project files.  It doesn't really make sense to keep this as part of the project when anytime you double click a .cs script file it rebuilds the solution for the UnityScreenApp itself.

Removing for now.  Might impact debugging ability, so might need to bring it back.  Otherwise it will work to have a second VS2017 launched with the C# code.