Rename and move scripts to make most sense.
Rename DrawSBS to LaunchAndPlay to more accurately describe what it does.  It's the primary game hooking and launching script.  Move this script to be active on the Player object, as the most logical place for this master controller.

Rename ScreenDistance to MoveScreen to describe what all it does now.  Not just distance, also up/down and resizing of screen are working.