Rework the RecenterHMD to handle room scale properly.
Changing the setup from RoomScale to Standing was a bit weird and has unknown side effects.  Rather than do that in order to Recenter, this just counter-rotates the Player from whereever the VRCamera is pointing, which is the way the HMD is facing.

Only counter rotates the Y direction to move the world to facing forward, and only does this once at launch or whenever the user choose F12 to recenter or Right Controller Grip.

This also adds the RightControllerGrip as a RecenterAction.

Tested and working.  No errors or warnings.  Grip works to recenter.