Add back GrabPinch and GrabGrip boolean actions.
With these missing, we are getting a null-pointer exception when bringing up the Dashboard.  Something to do with the Hand script requiring these to be available, otherwise the list of actions is null and it will crash.

Not super significant, just added the actions, but don't connect them with any binding.  This fixes the crash.