Counterrotate the floor so that it remains stable on RecenterHMD.
We rotate the world on RecenterHMD to get screen wherever the user wants to look. That rotates the Player object, which contains the SteamVR object which draws the world.  Because SteamVR rotates, the entire world rotates.

Counter rotate the floor itself, so that it appears to remain stable and matches their play space. It gives a good visual limit of the edge of their play space now, and having it rotate broke that visual cue.